moved compute_days_in_period to utils
